feat: Implement project modal functionality with dynamic loading and event handling
- Added project-modal.js to manage the opening, closing, and rendering of project modals.
- Integrated modal with project data, including title, description, status, type, and tech stack.
- Implemented image and Mermaid diagram display within the modal.
- Enhanced projects.js to utilize the new modal functions, including opening specific projects based on URL parameters.
- Refactored project card rendering to support modal interactions and improved error handling.
- Established event listeners for modal close actions and keyboard interactions.
- Loaded necessary templates dynamically for project cards and stack items.
